+++ src/index.d.ts
@@ -6,6 +6,24 @@ import {
 } from './services/config.js';
 
 import {
+	fetchArtistBySlug,
+	fetchArtistLessons,
+	fetchArtists
+} from './services/content/artist.ts';
+
+import {
+	fetchGenreBySlug,
+	fetchGenreLessons,
+	fetchGenres
+} from './services/content/genre.ts';
+
+import {
+	fetchInstructorBySlug,
+	fetchInstructorLessons,
+	fetchInstructors
+} from './services/content/instructor.ts';
+
+import {
 	enrollUserInGuidedCourse,
 	fetchEnrollmentPageMetadata,
 	guidedCourses,
@@ -13,10 +31,12 @@ import {
 } from './services/content-org/guided-courses.ts';
 
 import {
+	completeLearningPathIntroVideo,
+	completeMethodIntroVideo,
 	fetchLearningPathLessons,
 	getActivePath,
 	getDailySession,
-	getLearningPath,
+	getEnrichedLearningPath,
 	getLearningPathLessonsByIds,
 	mapContentToParent,
 	resetAllLearningPaths,
@@ -222,21 +242,18 @@ import {
 } from './services/recommendations.js';
 
 import {
+	buildEntityAndTotalQuery,
 	fetchAll,
 	fetchAllFilterOptions,
 	fetchAllPacks,
-	fetchArtistLessons,
-	fetchArtists,
 	fetchByRailContentId,
 	fetchByRailContentIds,
 	fetchByReference,
 	fetchChatAndLiveEnvent,
-	fetchCoachLessons,
 	fetchComingSoon,
 	fetchCommentModContentData,
 	fetchContentRows,
 	fetchFoundation,
-	fetchGenreLessons,
 	fetchHierarchy,
 	fetchLeaving,
 	fetchLessonContent,
@@ -272,6 +289,7 @@ import {
 	fetchTabData,
 	fetchTopLevelParentId,
 	fetchUpcomingEvents,
+	getSanityDate,
 	getSortOrder,
 	jumpToContinueContent
 } from './services/sanity.js';
@@ -404,9 +422,12 @@ declare module 'musora-content-services' {
 		assignModeratorToComment,
 		blockUser,
 		blockedUsers,
+		buildEntityAndTotalQuery,
 		buildImageSRC,
 		calculateLongestStreaks,
 		closeComment,
+		completeLearningPathIntroVideo,
+		completeMethodIntroVideo,
 		confirmEmailChange,
 		contentStatusCompleted,
 		contentStatusReset,
@@ -438,6 +459,7 @@ declare module 'musora-content-services' {
 		fetchAllCompletedStates,
 		fetchAllFilterOptions,
 		fetchAllPacks,
+		fetchArtistBySlug,
 		fetchArtistLessons,
 		fetchArtists,
 		fetchAwardsForUser,
@@ -448,7 +470,6 @@ declare module 'musora-content-services' {
 		fetchCertificate,
 		fetchChatAndLiveEnvent,
 		fetchChatSettings,
-		fetchCoachLessons,
 		fetchComingSoon,
 		fetchComment,
 		fetchCommentModContentData,
@@ -466,9 +487,14 @@ declare module 'musora-content-services' {
 		fetchFollowedThreads,
 		fetchForumCategories,
 		fetchFoundation,
+		fetchGenreBySlug,
 		fetchGenreLessons,
+		fetchGenres,
 		fetchHandler,
 		fetchHierarchy,
+		fetchInstructorBySlug,
+		fetchInstructorLessons,
+		fetchInstructors,
 		fetchInterests,
 		fetchLastInteractedChild,
 		fetchLatestThreads,
@@ -544,8 +570,8 @@ declare module 'musora-content-services' {
 		getAwardDataForGuidedContent,
 		getContentRows,
 		getDailySession,
+		getEnrichedLearningPath,
 		getLastInteractedOf,
-		getLearningPath,
 		getLearningPathLessonsByIds,
 		getLegacyMethods,
 		getLessonContentRows,
@@ -568,6 +594,7 @@ declare module 'musora-content-services' {
 		getRecommendedForYou,
 		getResumeTimeSeconds,
 		getResumeTimeSecondsByIds,
+		getSanityDate,
 		getScheduleContentRows,
 		getSortOrder,
 		getStartedOrCompletedProgressOnly,