+++ src/services/sanity.js
@@ -22,19 +22,16 @@ import {
   SONG_TYPES,
   SONG_TYPES_WITH_CHILDREN,
 } from '../contentTypeConfig.js'
-import {fetchSimilarItems, recommendations} from './recommendations.js'
+import { fetchSimilarItems, recommendations } from './recommendations.js'
 import { processMetadata, typeWithSortOrder } from '../contentMetaData.js'
 
 import { globalConfig } from './config.js'
 
-import {
-  fetchNextContentDataForParent,
-  fetchHandler,
-} from './railcontent.js'
+import { fetchNextContentDataForParent, fetchHandler } from './railcontent.js'
 import { arrayToStringRepresentation, FilterBuilder } from '../filterBuilder.js'
 import { fetchUserPermissions } from './user/permissions.js'
 import { getAllCompleted, getAllStarted, getAllStartedOrCompleted } from './contentProgress.js'
-import {fetchRecentActivitiesActiveTabs} from "./userActivity.js";
+import { fetchRecentActivitiesActiveTabs } from './userActivity.js'
 
 /**
  * Exported functions that are excluded from index generation.
@@ -160,30 +157,6 @@ function getQueryFromPage(pageNumber, contentPerPage) {
   result['start'] = start
   result['end'] = end
   return result
-}
-
-/**
- * Fetch all artists with lessons available for a specific brand.
- *
- * @param {string} brand - The brand for which to fetch artists.
- * @returns {Promise<Object|null>} - A promise that resolves to an array of artist objects or null if not found.
- *
- * @example
- * fetchArtists('drumeo')
- *   .then(artists => console.log(artists))
- *   .catch(error => console.error(error));
- */
-export async function fetchArtists(brand) {
-  const filter = await new FilterBuilder(
-    `_type == "song" && brand == "${brand}" && references(^._id)`,
-    { bypassPermissions: true }
-  ).buildFilter()
-  const query = `
-  *[_type == "artist"]{
-    name,
-    "lessonsCount": count(*[${filter}])
-  }[lessonsCount > 0] |order(lower(name)) `
-  return fetchSanity(query, true, { processNeedAccess: false })
 }
 
 /**
@@ -335,7 +308,7 @@ export async function fetchNewReleases(
       web_url_path,
       "permission_id": permission[]->railcontent_id,
       `
-  const query = buildRawQuery(filter, fields, {sortOrder: sortOrder, start, end: end})
+  const query = buildRawQuery(filter, fields, { sortOrder: sortOrder, start, end: end })
   return fetchSanity(query, true)
 }
 
@@ -476,17 +449,26 @@ export async function fetchByRailContentId(id, contentType) {
  *   .then(contents => console.log(contents))
  *   .catch(error => console.error(error));
  */
-export async function fetchByRailContentIds(ids, contentType = undefined, brand = undefined, includePermissionsAndStatusFilter = false) {
+export async function fetchByRailContentIds(
+  ids,
+  contentType = undefined,
+  brand = undefined,
+  includePermissionsAndStatusFilter = false
+) {
   if (!ids?.length) {
     return []
   }
-  ids = [...new Set(ids.filter(item => item !== null && item !== undefined))];
+  ids = [...new Set(ids.filter((item) => item !== null && item !== undefined))]
   const idsString = ids.join(',')
   const brandFilter = brand ? ` && brand == "${brand}"` : ''
-  const lessonCountFilter = await new FilterBuilder(`_id in ^.child[]._ref`, {pullFutureContent: true}).buildFilter()
+  const lessonCountFilter = await new FilterBuilder(`_id in ^.child[]._ref`, {
+    pullFutureContent: true,
+  }).buildFilter()
   const fields = await getFieldsForContentTypeWithFilteredChildren(contentType, true)
   const baseFilter = `railcontent_id in [${idsString}]${brandFilter}`
-  const finalFilter = includePermissionsAndStatusFilter ? await new FilterBuilder(baseFilter).buildFilter() : baseFilter
+  const finalFilter = includePermissionsAndStatusFilter
+    ? await new FilterBuilder(baseFilter).buildFilter()
+    : baseFilter
   const query = `*[
     ${finalFilter}
   ]{
@@ -498,18 +480,16 @@ export async function fetchByRailContentIds(ids, contentType = undefined, brand 
 
   console.log('ids query', query)
   const customPostProcess = (results) => {
-    const now = getSanityDate(new Date(), false);
+    const now = getSanityDate(new Date(), false)
     const liveProcess = (result) => {
       if (result.live_event_start_time && result.live_event_end_time) {
-        result.isLive =
-          result.live_event_start_time <= now &&
-          result.live_event_end_time >= now;
+        result.isLive = result.live_event_start_time <= now && result.live_event_end_time >= now
       } else {
-        result.isLive = false;
+        result.isLive = false
       }
-      return result;
-    };
-    return results.map(liveProcess);
+      return result
+    }
+    return results.map(liveProcess)
   }
   const results = await fetchSanity(query, true, { customPostProcess: customPostProcess })
 
@@ -527,13 +507,14 @@ export async function fetchByRailContentIds(ids, contentType = undefined, brand 
   return sortedResults
 }
 
-export async function fetchContentRows(brand, pageName, contentRowSlug)
-{
+export async function fetchContentRows(brand, pageName, contentRowSlug) {
   if (pageName === 'lessons') pageName = 'lesson'
   if (pageName === 'songs') pageName = 'song'
   const rowString = contentRowSlug ? ` && slug.current == "${contentRowSlug.toLowerCase()}"` : ''
-  const lessonCountFilter = await new FilterBuilder(`_id in ^.child[]._ref`, {pullFutureContent: true}).buildFilter()
-  const childFilter = await new FilterBuilder('', {isChildrenFilter: true}).buildFilter()
+  const lessonCountFilter = await new FilterBuilder(`_id in ^.child[]._ref`, {
+    pullFutureContent: true,
+  }).buildFilter()
+  const childFilter = await new FilterBuilder('', { isChildrenFilter: true }).buildFilter()
   const query = `*[_type == 'recommended-content-row' && brand == '${brand}' && type == '${pageName}'${rowString}]{
     brand,
     name,
@@ -548,8 +529,6 @@ export async function fetchContentRows(brand, pageName, contentRowSlug)
   }`
   return fetchSanity(query, true)
 }
-
-
 
 /**
  * Fetch all content for a specific brand and type with pagination, search, and grouping options.
@@ -619,9 +598,7 @@ export async function fetchAll(
   } else if (type === 'pack') {
     typeFilter = `&& (_type == 'pack' || _type == 'semester-pack')`
   } else {
-    typeFilter = type
-      ? `&& _type == '${type}'`
-      : ''
+    typeFilter = type ? `&& _type == '${type}'` : ''
   }
 
   // Construct the search filter
@@ -740,7 +717,7 @@ async function getProgressFilter(progress, progressIds) {
       return `&& (railcontent_id in [${ids.join(',')}])`
     }
     case 'incomplete': {
-      const ids = progressIds !== undefined ? progressIds :await getAllStarted()
+      const ids = progressIds !== undefined ? progressIds : await getAllStarted()
       return `&& railcontent_id in [${ids.join(',')}]`
     }
     default:
@@ -904,6 +881,7 @@ export async function fetchFoundation(slug) {
  * @param {string} slug - The slug of the method.
  * @returns {Promise<Object|null>} - The fetched methods data or null if not found.
  */
+//todo BEH-1446 depreciated. remove all old method functions
 export async function fetchMethod(brand, slug) {
   const childrenFilter = await new FilterBuilder(``, { isChildrenFilter: true }).buildFilter()
 
@@ -1247,7 +1225,7 @@ async function fetchRelatedByLicense(railcontentId, brand, onlyUseSongTypes, cou
           *[${filterSongTypesWithSameLicense}]->{${queryFields}}|order(published_on desc, title asc)[0...${count}],
       }[0...1]`
   const results = await fetchSanity(query, false)
-  return results ? results['related_by_license'] ?? [] : []
+  return results ? (results['related_by_license'] ?? []) : []
 }
 
 /**
@@ -1256,15 +1234,17 @@ async function fetchRelatedByLicense(railcontentId, brand, onlyUseSongTypes, cou
  * @param {string} brand - The current brand.
  * @returns {Promise<Array<Object>|null>} - The fetched related lessons data or null if not found.
  */
-export async function fetchSiblingContent(railContentId, brand= null)
-{
+export async function fetchSiblingContent(railContentId, brand = null) {
   const filterGetParent = await new FilterBuilder(`references(^._id) && _type == ^.parent_type`, {
-    pullFutureContent: true
-  }).buildFilter()
-  const filterForParentList = await new FilterBuilder(`references(^._id) && _type == ^.parent_type`, {
     pullFutureContent: true,
-    isParentFilter: true,
   }).buildFilter()
+  const filterForParentList = await new FilterBuilder(
+    `references(^._id) && _type == ^.parent_type`,
+    {
+      pullFutureContent: true,
+      isParentFilter: true,
+    }
+  ).buildFilter()
 
   const childrenFilter = await new FilterBuilder(``, { isChildrenFilter: true }).buildFilter()
 
@@ -1304,8 +1284,7 @@ export async function fetchSiblingContent(railContentId, brand= null)
  * @param {string} railContentId - The RailContent ID of the current lesson.
  * @returns {Promise<Array<Object>|null>} - The fetched related lessons data or null if not found.
  */
-export async function fetchRelatedLessons(railContentId)
-{
+export async function fetchRelatedLessons(railContentId) {
   const defaultFilterFields = `_type==^._type && brand == ^.brand && railcontent_id != ${railContentId}`
 
   const filterSameArtist = await new FilterBuilder(
@@ -1349,11 +1328,16 @@ export async function fetchAllPacks(
   const start = (page - 1) * limit
   const end = start + limit
 
-  const query = await buildQuery(filter, filterParams, await getFieldsForContentTypeWithFilteredChildren('pack'), {
-    sortOrder: sortOrder,
-    start,
-    end,
-  })
+  const query = await buildQuery(
+    filter,
+    filterParams,
+    await getFieldsForContentTypeWithFilteredChildren('pack'),
+    {
+      sortOrder: sortOrder,
+      start,
+      end,
+    }
+  )
   return fetchSanity(query, true)
 }
 
@@ -1461,45 +1445,6 @@ export async function fetchPackData(id) {
 
 /**
  * Fetch the data needed for the coach screen.
- * @param {string} brand - The brand for which to fetch coach lessons
- * @param {string} id - The Railcontent ID of the coach
- * @returns {Promise<Object|null>} - The lessons for the instructor or null if not found.
- * @param {Object} params - Parameters for pagination, filtering and sorting.
- * @param {string} [params.sortOrder="-published_on"] - The field to sort the lessons by.
- * @param {string} [params.searchTerm=""] - The search term to filter content by title.
- * @param {number} [params.page=1] - The page number for pagination.
- * @param {number} [params.limit=10] - The number of items per page.
- * @param {Array<string>} [params.includedFields=[]] - Additional filters to apply to the query in the format of a key,value array. eg. ['difficulty,Intermediate', 'genre,rock'].
- *
- * @example
- * fetchCoachLessons('coach123')
- *   .then(lessons => console.log(lessons))
- *   .catch(error => console.error(error));
- */
-export async function fetchCoachLessons(
-  brand,
-  id,
-  { sortOrder = '-published_on', searchTerm = '', page = 1, limit = 20, includedFields = [] } = {}
-) {
-  const fieldsString = getFieldsForContentType()
-  const start = (page - 1) * limit
-  const end = start + limit
-  const searchFilter = searchTerm ? `&& title match "${searchTerm}*"` : ''
-  const includedFieldsFilter = includedFields.length > 0 ? filtersToGroq(includedFields) : ''
-  const filter = `brand == '${brand}' ${searchFilter} ${includedFieldsFilter} && references(*[_type=='instructor' && railcontent_id == ${id}]._id)`
-  const filterWithRestrictions = await new FilterBuilder(filter).buildFilter()
-
-  sortOrder = getSortOrder(sortOrder, brand)
-  const query = buildEntityAndTotalQuery(filterWithRestrictions, fieldsString, {
-    sortOrder: sortOrder,
-    start: start,
-    end: end,
-  })
-  return fetchSanity(query, true)
-}
-
-/**
- * Fetch the data needed for the coach screen.
  * @param {string} id - The Railcontent ID of the coach
  *
  * @returns {Promise<Object|null>} - The lessons for the instructor or null if not found.
@@ -1526,121 +1471,6 @@ export async function fetchByReference(
     start: start,
     end: end,
   })
-  return fetchSanity(query, true)
-}
-
-/**
- * Fetch the artist's lessons.
- * @param {string} brand - The brand for which to fetch lessons.
- * @param {string} name - The name of the artist
- * @param {string} contentType - The type of the lessons we need to get from the artist. If not defined, groq will get lessons from all content types
- * @param {Object} params - Parameters for sorting, searching, pagination and filtering.
- * @param {string} [params.sort="-published_on"] - The field to sort the lessons by.
- * @param {string} [params.searchTerm=""] - The search term to filter the lessons.
- * @param {number} [params.page=1] - The page number for pagination.
- * @param {number} [params.limit=10] - The number of items per page.
- * @param {Array<string>} [params.includedFields=[]] - Additional filters to apply to the query in the format of a key,value array. eg. ['difficulty,Intermediate', 'genre,rock'].
- * @param {Array<number>} [params.progressIds] - The ids of the lessons that are in progress or completed
- * @returns {Promise<Object|null>} - The lessons for the artist and some details about the artist (name and thumbnail).
- *
- * @example
- * fetchArtistLessons('drumeo', '10 Years', 'song', {'-published_on', '', 1, 10, ["difficulty,Intermediate"], [232168, 232824, 303375, 232194, 393125]})
- *   .then(lessons => console.log(lessons))
- *   .catch(error => console.error(error));
- */
-export async function fetchArtistLessons(
-  brand,
-  name,
-  contentType,
-  {
-    sort = '-published_on',
-    searchTerm = '',
-    page = 1,
-    limit = 10,
-    includedFields = [],
-    progressIds = undefined,
-  } = {}
-) {
-  const fieldsString = DEFAULT_FIELDS.join(',')
-  const start = (page - 1) * limit
-  const end = start + limit
-  const searchFilter = searchTerm ? `&& title match "${searchTerm}*"` : ''
-  const sortOrder = getSortOrder(sort, brand)
-  const addType =
-    contentType && Array.isArray(contentType)
-      ? `_type in ['${contentType.join("', '")}'] &&`
-      : contentType
-        ? `_type == '${contentType}' && `
-        : ''
-  const includedFieldsFilter = includedFields.length > 0 ? filtersToGroq(includedFields) : ''
-
-  // limits the results to supplied progressIds for started & completed filters
-  const progressFilter =
-    progressIds !== undefined ? `&& railcontent_id in [${progressIds.join(',')}]` : ''
-  const now = getSanityDate(new Date())
-  const query = `{
-    "entity":
-      *[_type == 'artist' && name == '${name}']
-        {'type': _type, name, 'thumbnail':thumbnail_url.asset->url,
-        'lessons_count': count(*[${addType} brand == '${brand}' && references(^._id)]),
-        'lessons': *[${addType} brand == '${brand}' && references(^._id) && (status in ['published'] || (status == 'scheduled' && defined(published_on) && published_on >= '${now}')) ${searchFilter} ${includedFieldsFilter} ${progressFilter}]{${fieldsString}}
-      [${start}...${end}]}
-      |order(${sortOrder})f
-  }`
-  return fetchSanity(query, true)
-}
-
-/**
- * Fetch the genre's lessons.
- * @param {string} brand - The brand for which to fetch lessons.
- * @param {string} name - The name of the genre
- * @param {Object} params - Parameters for sorting, searching, pagination and filtering.
- * @param {string} [params.sort="-published_on"] - The field to sort the lessons by.
- * @param {string} [params.searchTerm=""] - The search term to filter the lessons.
- * @param {number} [params.page=1] - The page number for pagination.
- * @param {number} [params.limit=10] - The number of items per page.
- * @param {Array<string>} [params.includedFields=[]] - Additional filters to apply to the query in the format of a key,value array. eg. ['difficulty,Intermediate', 'genre,rock'].
- * @param {Array<number>} [params.progressIds] - The ids of the lessons that are in progress or completed
- * @returns {Promise<Object|null>} - The lessons for the artist and some details about the artist (name and thumbnail).
- *
- * @example
- * fetchGenreLessons('drumeo', 'Blues', 'song', {'-published_on', '', 1, 10, ["difficulty,Intermediate"], [232168, 232824, 303375, 232194, 393125]})
- *   .then(lessons => console.log(lessons))
- *   .catch(error => console.error(error));
- */
-export async function fetchGenreLessons(
-  brand,
-  name,
-  contentType,
-  {
-    sort = '-published_on',
-    searchTerm = '',
-    page = 1,
-    limit = 10,
-    includedFields = [],
-    progressIds = undefined,
-  } = {}
-) {
-  const fieldsString = DEFAULT_FIELDS.join(',')
-  const start = (page - 1) * limit
-  const end = start + limit
-  const searchFilter = searchTerm ? `&& title match "${searchTerm}*"` : ''
-  const sortOrder = getSortOrder(sort, brand)
-  const addType = contentType ? `_type == '${contentType}' && ` : ''
-  const includedFieldsFilter = includedFields.length > 0 ? filtersToGroq(includedFields) : ''
-  // limits the results to supplied progressIds for started & completed filters
-  const progressFilter =
-    progressIds !== undefined ? `&& railcontent_id in [${progressIds.join(',')}]` : ''
-  const now = getSanityDate(new Date())
-  const query = `{
-    "entity":
-      *[_type == 'genre' && name == '${name}']
-        {'type': _type, name, 'thumbnail':thumbnail_url.asset->url,
-        'lessons_count': count(*[${addType} brand == '${brand}' && references(^._id)]),
-        'lessons': *[${addType} brand == '${brand}' && references(^._id) && (status in ['published'] || (status == 'scheduled' && defined(published_on) && published_on >= '${now}')) ${searchFilter} ${includedFieldsFilter} ${progressFilter}]{${fieldsString}}
-      [${start}...${end}]}
-      |order(${sortOrder})
-  }`
   return fetchSanity(query, true)
 }
 
@@ -1765,8 +1595,10 @@ export async function fetchCommentModContentData(ids) {
  *
  * @param {string} query - The GROQ query to execute against the Sanity API.
  * @param {boolean} isList - Whether to return an array or a single result.
- * @param {Function} [customPostProcess=null] - custom post process callback
- * @param {boolean} [processNeedAccess=true] - execute the needs_access callback
+ * @param {Object} options - Additional options for fetching data.
+ * @param {Function} [options.customPostProcess=null] - custom post process callback
+ * @param {boolean} [options.processNeedAccess=true] - execute the needs_access callback
+ * @param {boolean} [options.processPageType=true] - execute the page_type callback
  * @returns {Promise<*|null>} - A promise that resolves to the fetched data or null if an error occurs or no results are found.
  *
  * @example
@@ -1779,7 +1611,7 @@ export async function fetchCommentModContentData(ids) {
 export async function fetchSanity(
   query,
   isList,
-  { customPostProcess = null, processNeedAccess = true } = {}
+  { customPostProcess = null, processNeedAccess = true, processPageType = true } = {}
 ) {
   // Check the config object before proceeding
   if (!checkSanityConfig(globalConfig)) {
@@ -1821,19 +1653,20 @@ export async function fetchSanity(
       results = processNeedAccess
         ? await needsAccessDecorator(results, userPermissions, isAdmin)
         : results
-      results = pageTypeDecorator(results)
+      results = processPageType
+        ? pageTypeDecorator(results)
+        : results
       return customPostProcess ? customPostProcess(results) : results
     } else {
       throw new Error('No results found')
     }
   } catch (error) {
-    console.error('fetchSanity: Fetch error:', {error, query})
+    console.error('fetchSanity: Fetch error:', { error, query })
     return null
   }
 }
 
-function contentResultsDecorator(results, fieldName, callback)
-{
+function contentResultsDecorator(results, fieldName, callback) {
   if (Array.isArray(results)) {
     results.forEach((result) => {
       result[fieldName] = callback(result)
@@ -1860,16 +1693,18 @@ function contentResultsDecorator(results, fieldName, callback)
   return results
 }
 
-function pageTypeDecorator(results)
-{
-  return contentResultsDecorator(results, 'page_type', function(content) { return SONG_TYPES_WITH_CHILDREN.includes(content['type']) ? 'song' : 'lesson'})
+function pageTypeDecorator(results) {
+  return contentResultsDecorator(results, 'page_type', function (content) {
+    return SONG_TYPES_WITH_CHILDREN.includes(content['type']) ? 'song' : 'lesson'
+  })
 }
-
 
 function needsAccessDecorator(results, userPermissions, isAdmin) {
   if (globalConfig.sanityConfig.useDummyRailContentMethods) return results
   userPermissions = new Set(userPermissions)
-  return contentResultsDecorator(results, 'need_access', function (content) { return doesUserNeedAccessToContent(content, userPermissions, isAdmin) })
+  return contentResultsDecorator(results, 'need_access', function (content) {
+    return doesUserNeedAccessToContent(content, userPermissions, isAdmin)
+  })
 }
 
 function doesUserNeedAccessToContent(result, userPermissions, isAdmin) {
@@ -1927,8 +1762,8 @@ export async function fetchShowsData(brand) {
  *   .catch(error => console.error(error));
  */
 export async function fetchMetadata(brand, type) {
-  let processedData =  processMetadata(brand, type, true)
-  if(processedData?.onlyAvailableTabs === true) {
+  let processedData = processMetadata(brand, type, true)
+  if (processedData?.onlyAvailableTabs === true) {
     const activeTabs = await fetchRecentActivitiesActiveTabs()
     processedData.tabs = activeTabs
   }
@@ -1954,7 +1789,7 @@ function arrayJoinWithQuotes(array, delimiter = ',') {
   return wrapped.join(delimiter)
 }
 
-function getSanityDate(date, roundToHourForCaching = true) {
+export function getSanityDate(date, roundToHourForCaching = true) {
   if (roundToHourForCaching) {
     // We need to set the published on filter date to be a round time so that it doesn't bypass the query cache
     // with every request by changing the filter date every second. I've set it to one minute past the current hour
@@ -2028,13 +1863,19 @@ async function buildQuery(
   return buildRawQuery(filter, fields, { sortOrder, start, end, isSingle })
 }
 
-function buildEntityAndTotalQuery(
+export function buildEntityAndTotalQuery(
   filter = '',
   fields = '...',
-  { sortOrder = 'published_on desc', start = 0, end = 10, isSingle = false, withoutPagination = false }
+  {
+    sortOrder = 'published_on desc',
+    start = 0,
+    end = 10,
+    isSingle = false,
+    withoutPagination = false,
+  }
 ) {
   const sortString = sortOrder ? ` | order(${sortOrder})` : ''
-  const countString = isSingle ? '[0...1]' : (withoutPagination ? ``: `[${start}...${end}]`)
+  const countString = isSingle ? '[0...1]' : withoutPagination ? `` : `[${start}...${end}]`
   const query = `{
       "entity": *[${filter}]  ${sortString}${countString}
       {
@@ -2163,22 +2004,22 @@ export async function fetchTabData(
 
   switch (progress) {
     case 'recent':
-      progressIds = await getAllStartedOrCompleted({ brand, onlyIds: true });
-      sortOrder = null;
-      break;
+      progressIds = await getAllStartedOrCompleted({ brand, onlyIds: true })
+      sortOrder = null
+      break
     case 'incomplete':
-      progressIds = await getAllStarted();
-      sortOrder = null;
-      break;
+      progressIds = await getAllStarted()
+      sortOrder = null
+      break
     case 'completed':
-      progressIds = await getAllCompleted();
-      sortOrder = null;
-      break;
+      progressIds = await getAllCompleted()
+      sortOrder = null
+      break
   }
 
   // limits the results to supplied progressIds for started & completed filters
   const progressFilter = await getProgressFilter(progress, progressIds)
-  const fieldsString = getFieldsForContentType('tab-data');
+  const fieldsString = getFieldsForContentType('tab-data')
   const now = getSanityDate(new Date())
 
   // Determine the group by clause
@@ -2190,8 +2031,7 @@ export async function fetchTabData(
   const childrenFilter = await new FilterBuilder(``, { isChildrenFilter: true }).buildFilter()
   const childrenFields = await getChildFieldsForContentType('tab-data')
   const lessonCountFilter = await new FilterBuilder(`_id in ^.child[]._ref`).buildFilter()
-  entityFieldsString =
-    ` ${fieldsString}
+  entityFieldsString = ` ${fieldsString}
     'children': child[${childrenFilter}]->{ ${childrenFields} 'children': child[${childrenFilter}]->{ ${childrenFields} }, },
     'isLive': live_event_start_time <= "${now}" && live_event_end_time >= "${now}",
     'lesson_count': coalesce(count(*[${lessonCountFilter}]), 0),
@@ -2207,23 +2047,23 @@ export async function fetchTabData(
   query = buildEntityAndTotalQuery(filterWithRestrictions, entityFieldsString, {
     sortOrder: sortOrder,
     start: start,
-    end: end
+    end: end,
   })
 
-  let results = await fetchSanity(query, true);
+  let results = await fetchSanity(query, true)
 
   if (['recent', 'incomplete', 'completed'].includes(progress) && results.entity.length > 1) {
     const orderMap = new Map(progressIds.map((id, index) => [id, index]))
     results.entity = results.entity
       .sort((a, b) => {
-        const aIdx = orderMap.get(a.id) ?? Number.MAX_SAFE_INTEGER;
-        const bIdx = orderMap.get(b.id) ?? Number.MAX_SAFE_INTEGER;
-        return aIdx - bIdx || new Date(b.published_on) - new Date(a.published_on);
+        const aIdx = orderMap.get(a.id) ?? Number.MAX_SAFE_INTEGER
+        const bIdx = orderMap.get(b.id) ?? Number.MAX_SAFE_INTEGER
+        return aIdx - bIdx || new Date(b.published_on) - new Date(a.published_on)
       })
-      .slice(start, end);
+      .slice(start, end)
   }
 
-  return results;
+  return results
 }
 
 export async function fetchRecent(
@@ -2298,12 +2138,12 @@ export async function fetchShows(brand, type, sort = 'sort') {
  * @returns {Promise<*|null>}
  */
 export async function fetchMethodV2IntroVideo(brand) {
-  const type = "method-intro";
-  const filter = `_type == '${type}' && brand == '${brand}'`;
-  const fields = getIntroVideoFields('method-v2');
+  const type = 'method-intro'
+  const filter = `_type == '${type}' && brand == '${brand}'`
+  const fields = getIntroVideoFields('method-v2')
 
-  const query = `*[${filter}] { ${fields.join(", ")} }`;
-  return fetchSanity(query, false);
+  const query = `*[${filter}] { ${fields.join(', ')} }`
+  return fetchSanity(query, false)
 }
 
 /**
@@ -2312,7 +2152,7 @@ export async function fetchMethodV2IntroVideo(brand) {
  * @returns {Promise<*|null>}
  */
 export async function fetchMethodV2Structure(brand) {
-  const _type = "method-v2";
+  const _type = 'method-v2'
   const query = `*[_type == '${_type}' && brand == '${brand}'][0...1]{
     'sanity_id': _id,
     'learningPaths': child[]->{
@@ -2320,7 +2160,7 @@ export async function fetchMethodV2Structure(brand) {
       'children': child[]->railcontent_id
     }
   }`
-  return await fetchSanity(query, false);
+  return await fetchSanity(query, false)
 }
 
 /**